+++ airbnb-app/src/App.css
@@ -101,8 +101,10 @@ body {
 
 .cards .card h2{
     padding: 20px;
+    font-size: 16px;
     font-weight: 600;
     text-align: center;
+    
 }
